Labs/Lab01/requirements.txt
streamlit 
azure-storage-blob 
pyodbc
python-dotenv